(fix): TypeScript autocomplete
