+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 .vscode
 .DS_STORE
 coverage
+.output
