+++ src/module.ts
@@ -47,6 +47,11 @@ export default defineNuxtModule<ModuleOptions>({
         defaultUserAgent: options.defaultUserAgent,
         refreshOnResize: options.refreshOnResize
       })
+
+      nuxt.hook('prepare:types', (options) => {
+        options.references.push({ path: resolve(nuxt.options.buildDir, 'runtime/generateFlags.d.ts') })
+      })
+
       const runtimeDir = fileURLToPath(new URL('./runtime', import.meta.url))
       nuxt.options.build.transpile.push(runtimeDir)
       addPlugin(resolve(runtimeDir, 'plugin'))
